requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# GraphQL
strawberry-graphql[fastapi]>=0.217.0

# Database ORM & Driver
sqlmodel>=0.0.14
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0

# HTTP Client
httpx>=0.26.0

# Data Processing
polars>=0.20.0

# Environment & Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
celery[redis]>=5.3.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
